🎨 style(Footer.vue): remove unnecessary code and add placeholder text The footer background image and social media links have been removed as they are no longer needed. A placeholder text has been added to the footer to indicate that it is present.
🎉 feat(NavItems.vue): add new file
A new file, NavItems.vue, has been added to the project. This file will contain the navigation items for the application.

🎨 style(Navbar.vue): refactor Navbar component to improve readability and accessibility
The Navbar component has been refactored to improve its readability and accessibility. The changes include replacing the previous navigation bar with a new one that is more accessible and easier to use. The new navigation bar includes dropdown menus for Clinics and Diagnostics, which can be expanded or collapsed as needed. The new navigation bar also includes a Try for Free button and a Log In button for improved user experience.

🎨 style(Main.vue): add data-theme attribute to div elements
The data-theme attribute has been added to the div elements to allow for dynamic theming of the application. The div with data-theme="light" sets the default theme to light, while the div with data-theme="mycure" sets the theme to mycure.

✨ feat(index.vue): add new sections and components to the index page
This commit adds new sections and components to the index page. The first section is a hero section that includes a title and subtitle. It also includes a button to start for free and an image. The second section is a grid of clinic names. The third section is another hero section that includes a title and subtitle. The fourth section is a grid of features that includes an image, title, and description. The fifth section is a panel with a blue gradient background that includes a title, subtitle, and button. It also includes a grid of hello MYCURE text. The last section is a panel that includes a title and buttons with different styles.