🔧 chore(firebase): update firebase hosting configuration The `.firebaserc` file is added with the default project set to `mc-v4-prod`. The `firebase-hosting-merge.yml` and `firebase-hosting-pull-request.yml` files are updated to use the `mc-v4-prod` project ID and the `FIREBASE_SERVICE_ACCOUNT_MC_V4_PROD` secret. The `firebase-hosting-pull-request.yml` file is also updated to include the `channelId` and `expires` options. The `yarn generate` command is now run directly instead of installing dependencies and running the command separately.
🆕 chore(firebase.json): add firebase hosting configuration
🔧 chore(package.json): add @nuxt/devtools as a dev dependency
The firebase.json file has been added to the repository, which contains the configuration for Firebase hosting. The hosting configuration specifies the site name, public directory, and files to ignore. The package.json file has been updated to include @nuxt/devtools as a dev dependency. This package provides a browser extension for debugging Nuxt.js applications.