🎨 style(Footer.vue): refactor footer component to use GenericPanel component The Footer component has been refactored to use the GenericPanel component. This improves consistency and reduces code duplication. The footer now contains a logo and links to various sections of the website.
🎨 style(GenericPanel.vue): adjust max width of the panel on medium screens
The max width of the panel was adjusted to be 100% on medium screens, which improves the layout of the component.

🎨 style(Navbar.vue): refactor Navbar component to use generic-panel component
💄 style(Navbar.vue): update Navbar layout and design
The Navbar component has been refactored to use the generic-panel component, which improves code reusability and reduces code duplication. The layout and design of the Navbar have also been updated to improve the user experience. The changes include adding a logo, updating the dropdown menu design, and updating the clinic and diagnostics icons.

🎨 style(Navbar.vue): adjust Diagnostics icon size
The Diagnostics icon size has been adjusted to match the size of the other icons in the Navbar.

🎨 style(Navbar.vue): refactor navbar-end section
The navbar-end section has been refactored to use smaller buttons on medium-sized screens. The Log In and Try for Free buttons now have the btn-sm class on small screens and the btn-md class on medium-sized screens. The navbar-end section has also been moved inside the generic-panel component.

🎨 style(index.vue): adjust margins and padding of generic-panel components
The margins and padding of the generic-panel components have been adjusted to improve the layout of the page. The top margin of the first generic-panel component has been increased to 32px, while the bottom margin has been decreased to 20px. The second generic-panel component has been given a bottom margin of 32px and a padding of 12px on the y-axis.

🎨 style(index.vue): update panel and text styles
✨ feat(index.vue): add new sections and content
This commit updates the styles of the panels and text in the index.vue file. It also adds new sections and content to the page, including a new panel with a video, a section on comprehensive security, and a section featuring testimonials from satisfied customers. The changes aim to improve the user experience and provide more information about MYCURE's features and benefits.

🎨 style(index.vue): add search-card class with linear gradient background
This commit adds a new CSS class called search-card with a linear gradient background. The gradient is defined using the colors #004770 and rgba(0, 35, 88, 0.45) and is applied to the search card element. This improves the visual appeal of the search card and makes it stand out more on the page.