🔥 refactor(signup): remove trial property from form data The trial property was removed from the form data in both the health-facilities/index.vue and health-facilities/pricing.vue files. This property was not needed and was causing issues with the form submission.
🔨 refactor(axios/index.js): improve code readability by adding line breaks and spaces
This commit improves the readability of the code by adding line breaks and spaces to make it more readable. The changes do not affect the functionality of the code.

🔨 refactor(axios/index.js): simplify fields array declaration
🚀 feat(axios/index.js): add object destructuring to improve code readability
The fields array declaration is simplified by removing unnecessary line breaks and spaces. Object destructuring is used to improve code readability by making it easier to understand the structure of the payload object.